pyproject.toml
[tool.poetry]
name = "recommender_system"
version = "0.1.0"
description = ""
authors = ["Tom Matthews <tomukmatthews@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "1.18.1"
pandas = "1.1.3"
streamlit = "0.71.0"
tmdbsimple = "2.7.0"
tqdm = "4.55.0"
scikit-learn = "0.24.1"
requests = "2.25.0"
Pillow = "8.0.1"
pytest = "^6.2.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 100
